Conformed the ws example to the other examples. Corrected to work.
