+++ examples/websocket/html/websocket.html
